backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
aiomysql==0.2.0
aiosqlite==0.19.0
sqlalchemy==2.0.23
alembic==1.12.1
greenlet==3.0.1

# Vector database
pymilvus==2.3.4

# ML and NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.4

# HTTP clients and web scraping
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1